requirements.txt
allure-pytest==2.6.3
allure-python-commons==2.6.3
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrdict==2.0.1
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cheese==0.1.0
Click==7.0
coverage==4.5.3
cryptography==2.6.1
enum34==1.1.6
execnet==1.6.0
idna==2.8
json-tools==0.4.1
jsondiff==1.1.2
lxml==4.3.3
more-itertools==7.0.0
namedlist==1.7
paho-mqtt==1.4.0
paramiko==2.4.2
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pyfiglet==0.7.5
pymongo==3.8.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pynsq==0.8.3
pytest==4.5.0
pytest-cache==1.0
pytest-cov==2.7.1
pytest-mock==1.10.4
redis==3.2.1
requests==2.21.0
requests-mock==1.6.0
requests-toolbelt==0.9.1
six==1.12.0
tinydb==3.13.0
tornado==4.5.3
urllib3==1.24.3
wcwidth==0.1.7
timeout-decorator==0.4.1
pyyaml==5.3.1
grpcio==1.31.0
grpcio-tools==1.31.0
protobuf==3.13.0
colorama==0.4.3
